fixing errors and add small changes
