+++ src/main/java/org/example/controllers/ProfessorController.java
@@ -1,6 +1,9 @@
-package org.example;
+package org.example.controllers;
 
 
+import org.example.Professor;
+import org.example.repositories.ProfessorRepository;
+import org.example.Project;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
