+++ src/main/java/org/example/repositories/StudentRepository.java
@@ -1,12 +1,10 @@
-package org.example;
+package org.example.repositories;
 
+import org.example.Student;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-import java.util.ArrayList;
 import java.util.List;
 
-@RepositoryRestResource
 public interface StudentRepository extends CrudRepository<Student, Long> {
     @Override
     List<Student> findAll();
