+++ src/main/resources/templates/project.html
@@ -32,9 +32,6 @@
     <label>Description:</label>
     <input type="text" th:field="*{description}" required/><br/>
 
-    <label>Supervisor:</label>
-    <input type="text" th:field="*{professor}" /><br/>
-
     <label>Capacity:</label>
     <input type="number" th:field="*{capacity}" required/><br/>
