+++ src/main/resources/templates/student.html
@@ -9,17 +9,15 @@
     </style>
 </head>
 <body>
-<h2>Add Student</h2> // overview to add a student
-<input type="number" id="studentID" placeholder="Student ID"> // enter the student's ID
-<input type="text" id="studentName" placeholder="Student Name"> // enter the student's name
-<input type="text" id="studentEmail" placeholder="Student Email"> // enter the student's email
+<h2>Add Student</h2>
+<input type="number" id="studentID" placeholder="Student ID">
+<input type="text" id="studentName" placeholder="Student Name">
+<input type="text" id="studentEmail" placeholder="Student Email">
 <button onclick="addStudent()">Add Student</button>
-// add button to add the student
 
-<h2>Remove Student</h2> // overview to remove a student
-<input type="number" id="removeID" placeholder="Student ID"> // enter the student's ID
+<h2>Remove Student</h2>
+<input type="number" id="removeID" placeholder="Student ID">
 <button onclick="removeStudent()">Remove Student</button>
-// remove button to remove the student
 
 <h3>Current Students</h3>
 <ul id="studentList"></ul>
